anova_ana {MetENP}	R Documentation

Performs anova analysis Performs an anova across all response variables. The function takes 2 independent variables.

Description

Performs anova analysis Performs an anova across all response variables. The function takes 2 independent variables.

Usage

anova_ana(
  metabolomics_data,
  met_col,
  analysis_type,
  sample_col,
  metadata,
  normalization,
  factor_col1,
  factor_col2,
  p_adjust
)

Arguments

metabolomics_data
metabolomics data

met_col
column with metabolite names

analysis_type
type of analysis ex-GCMS, HILIAC positive ion mode.

sample_col
the column name having samples

metadata
Metadata

normalization
normalization methods any of the three: half_of_min, remove_NAs and 50percent

half_of_min: where the NAs are replaced by half of min values in the data

remove_NAs: where Cols with NAs values are removed

50percent: where cols with more than 50% NAs values are removed

factor_col1
column name of first independent variable

factor_col2
column name of second independent variable

p_adjust
Method for p value adjustment, i.e. "fdr"

Examples

anova_result = anova_ana(metabolomics_data=data,met_col="metabolite_name",analysis_type=c('Phospholipids, Chol. esters and Diacylglycerols','Sphingolipids'), metadata=metadata,normalization="50percent", factor_col1="TreatmentGroup",factor_col2="SamplingTimePoint",p_adjust="fdr")

[Package MetENP version 0.0.0.9000 Index]