compoundinfo {MetENP}	R Documentation

Counts total no of compound in a pathway that has been linked to significant metabolites This will help in calculating HG score for metabolite enrichment

Description

Counts total no of compound in a pathway that has been linked to significant metabolites This will help in calculating HG score for metabolite enrichment

Usage

compoundinfo(met_path, sps)

Arguments

met_path
dataframe of pathways mapped to significant metabolites as obtained from met_pathways

sps
species name

Examples

compound_count = compoundinfo(met_path, sps)

[Package MetENP version 0.0.0.9000 Index]