getmwstudies {MetENP}	R Documentation

Gets metabolomics data, metadata and metabolite info from Metabolomics Workbench using REST service

Description

Gets metabolomics data, metadata and metabolite info from Metabolomics Workbench using REST service

Usage

getmwstudies(studyid, typedata)

Arguments

studyid
from MW and typedata whether metadata (factors) or metabolic data

typedata
can be "factors" for metadata, "data" for metabolomics data and 'metabolites' for metabolite info

Examples

data = getmwstudies('ST001304', 'data')

[Package MetENP version 0.0.0.9000 Index]