mapspspath {MetENP}	R Documentation

Maps species specific pathways This is important because sometimes some species may not have some pathways

Description

Maps species specific pathways This is important because sometimes some species may not have some pathways

Usage

mapspspath(sps, rxn_pathway_df)

Arguments

sps
species name

rxn_pathway_df
dataframe having pathway information on metabolites as obtained from pathinfo

Examples

org_pathways_map= mapspspath(sps=sps,rxn_pathway_df= res_path)

[Package MetENP version 0.0.0.9000 Index]