met_pathways {MetENP}	R Documentation

Gives information on pathways linked to significant metabolites specific to species of choice

Description

Gives information on pathways linked to significant metabolites specific to species of choice

Usage

met_pathways(df_metenrichment, sps)

Arguments

df_metenrichment
metenrichment score dataframe having metabolites, metabolite class and their enrichment score obtained from metclassenrichment

sps
species name

Examples

met_path = met_pathways(df_metenrichment = metenrichment, 'hsa')

[Package MetENP version 0.0.0.9000 Index]