metclassenrichment {MetENP}	R Documentation

Provides enrichment score of the selected metabolic class via hypergeometric score

Description

Provides enrichment score of the selected metabolic class via hypergeometric score

Usage

metclassenrichment(df_metclass, refmet_class, metclass, enrich_stats, no)

Arguments

df_metclass
dataframe having significant metabolite with metabolite class and kegg ids

refmet_class
dataframe having all the metabolites in the study with refmet classification

metclass
Sub class, Main class or Super class

enrich_stats
HG for hypergeometric score ## leaves room for further including of KS stats

no
number of significant metabolites that should be in a class, default 1, should increase to 3 or more for a better statistically significant score. Default is chosen as 1 so as to get information on all the significant metabolites without any filtering.

Examples

metenrichment = metclassenrichment(df_metclass=sig_metabolites_kegg_id, refmet_class, metclass= "Sub class",enrich_stats="HG",no=1)

[Package MetENP version 0.0.0.9000 Index]