path_enrichmentscore {MetENP}	R Documentation

Calculates pathway enrichment score of the pathways linked to significant metabolites

Description

Calculates pathway enrichment score of the pathways linked to significant metabolites

Usage

path_enrichmentscore(
  met_path,
  sig_metabolite_kegg_id,
  ls_path,
  refmet_class,
  sps,
  padj,
  kegg_comp_path = FALSE
)

Arguments

met_path
dataframe of pathways mapped to significant metabolites as obtained from met_pathways

sig_metabolite_kegg_id
dataframe of significantly altered metabolites for HG calculation

ls_path
list of all the KEGG compounds linked to pathways. This can be obtained by running comp_linkedto_pathways().R or is also available please use ls_path when you want to calculate HG score depending on all the compounds in KEGG linked to a pathway sysdata.rda

refmet_class
dataframe of metabolite data with refmet class for a study, alternatively, data obtained from rest can also be given here which is yet not associated to refmet class We need this to get the number of all the metabolites detected in the study for calculating HG score.

sps
species name

padj
Method for p value adjustment, i.e. "fdr"

kegg_comp_path
TRUE or FALSE depending you want to use the ls_path as your background. Default is FALSE, where background set is the total metabolite in a study

Examples

kegg_es = path_enrichmentscore(met_path,sig_metabolite_kegg_id, ls_path,refmet_class,sps='hsa',padj='fdr')

[Package MetENP version 0.0.0.9000 Index]