pathinfo {MetENP}	R Documentation

Provides pathway or orthology information from reaction data of significant metabolite

Description

Provides pathway or orthology information from reaction data of significant metabolite

Usage

pathinfo(res, col_name)

Arguments

res
dataframe having metabolite information along with reaction information as obtained from rxninfo

col_name
"PATHWAY" or "ORTHOLOGY"

Examples

res_path= pathinfo(res, 'PATHWAY') ###rxn pathway

[Package MetENP version 0.0.0.9000 Index]