plot_met_enrichment {MetENP}	R Documentation

Plots metabolite enrichment score

Description

Plots metabolite enrichment score

Usage

plot_met_enrichment(df_metenrichment, metclass, enrich_stats, no)

Arguments

df_metenrichment
metenrichment score dataframe having metabolites, metabolite class and their enrichment score obtained from metclassenrichment

metclass
= sub_class, main_class or super_class

enrich_stats
HG for hypergeometric score ## leaves room for further including of KS stats

no
number of significant metabolites that should be in a class, default 1, should increase to 3 or more for a better statistically significant score. Default is chosen as 1 so as to get information on all the significant metabolites without any filtering.

Examples

plot_met_enrichment(metenrichment, "sub_class","HG", no=1)

[Package MetENP version 0.0.0.9000 Index]