plot_pathway_networks {MetENP}	R Documentation

Plots 1st degree pathway network of only pathways having significant pathway score

Description

Plots 1st degree pathway network of only pathways having significant pathway score

Usage

plot_pathway_networks(met_path, kegg_es, colorEdge)

Arguments

met_path
dataframe of pathways mapped to significant metabolites

kegg_es
dataframe of enrichment score of pathways

colorEdge
default TRUE to color the edges by pathways

Examples

plot_pathway_networks (met_path,kegg_es, TRUE)

[Package MetENP version 0.0.0.9000 Index]