react_substrate {MetENP}	R Documentation

Gets information whether metabolite is a product or substrate

Description

Gets information whether metabolite is a product or substrate

Usage

react_substrate(met_gene_info)

Arguments

met_gene_info
dataframe with gene information obtained from enzyme_gene_info

Examples

rclass_info = react_substrate(met_gene_info)

[Package MetENP version 0.0.0.9000 Index]