separate_data {MetENP}	R Documentation

Separates user data into metadata and metabolomics data with metabolite measurements

Description

Separates user data into metadata and metabolomics data with metabolite measurements

Usage

separate_data(example, typedata, sample_col)

Arguments

example
user data file

typedata
type of data desired from the user data uploaded: metadata or metabolomics data

sample_col
TRUE or FALSE, TRUE if sample names are in the first column, FALSE if metabolites names are in the first column.

Examples

separate_data('example.txt', "metadata",FALSE)

[Package MetENP version 0.0.0.9000 Index]